requirements.txt
fastapi
uvicorn
pymupdf>=1.18.0
pillow
python-multipart
groq
python-dotenv
inference_sdk
tensorflow